+++ R/modelsummary.R
@@ -85,7 +85,8 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low',
 #' define the labels that must appear in the table, and its names identify the
 #' original term names stored in the model object: `c("hp:mpg"="HPxM/G")`. See
 #' Examples section below.
-#' @param coef_omit string regular expression (perl-compatible) used to determine which coefficients to omit from the table. A "negative lookahead" can be used to specify which coefficients to *keep* in the table. Examples:
+#' @param coef_omit integer vector or string regular expression (perl-compatible) used to determine which coefficients to omit from the table. Note: A "negative lookahead" can be used to specify which coefficients to *keep* in the table. Examples:
+#' * c(2, 3, 5): omits the second, third, and fifth coefficients.
 #' * `"ei"`: omit coefficients matching the "ei" substring.
 #' * `"^Volume$"`: omit the "Volume" coefficient.
 #' * `"ei|rc"`: omit coefficients matching either the "ei" or the "rc" substrings.
@@ -332,6 +333,7 @@ modelsummary <- function(
     }
   }
 
+
   est <- shape_estimates(est, shape, conf_level = conf_level)
 
   # distinguish between estimates and gof (first column for tests)
@@ -369,6 +371,21 @@ modelsummary <- function(
   }
 
   est <- est[do.call(order, as.list(est)), ]
+
+  if (is.numeric(coef_omit)) {
+    coef_omit <- unique(round(coef_omit))
+    if (!"term" %in% shape$lhs) {
+      msg <- "`term` must be on the left-hand side of the `shape` formula when `coef_omit` is a numeric vector."
+      insight::format_error(msg)
+    }
+    term_idx <- paste(est$group, est$term)
+    if (max(coef_omit) > length(unique(term_idx))) {
+      msg <- sprintf("There are %s unique terms, but `coef_omit` tried to omit more than that.", length(term_idx))
+      insight::format_error(msg)
+    }
+    idx <- !term_idx %in% unique(term_idx)[coef_omit]
+    est <- est[idx, , drop = FALSE]
+  }
 
   # we kept the group column until here for sorting of mixed-effects by group
   if (is.null(shape$group_name)) {