+++ man-roxygen/modelsummary_examples.R
@@ -76,20 +76,23 @@
 #' modelsummary(models, coef_map = c('Volume' = 'Large', 'Height' = 'Tall'))
 #' modelsummary(models, coef_map = c('Volume', 'Height'))
 #'
+#' # coef_omit: omit the first and second coefficients
+#' modelsummary(models, coef_omit = 1:2)
+#' 
 #' # coef_omit: omit coefficients matching one substring
-#' modelsummary(models, coef_omit = "ei", omit = ".*")
+#' modelsummary(models, coef_omit = "ei", gof_omit = ".*")
 #'
 #' # coef_omit: omit a specific coefficient
 #' modelsummary(models, coef_omit = "^Volume$", gof_omit = ".*")
 #' 
 #' # coef_omit: omit coefficients matching either one of two substring
-#' modelsummary(models, coef_omit = "ei|rc", omit = ".*")
+#' modelsummary(models, coef_omit = "ei|rc", gof_omit = ".*")
 #' 
 #' # coef_omit: keep coefficients starting with a substring (using a negative lookahead)
-#' modelsummary(models, coef_omit = "^(?!Vol)", omit = ".*")
+#' modelsummary(models, coef_omit = "^(?!Vol)", gof_omit = ".*")
 #' 
 #' # coef_omit: keep coefficients matching a substring
-#' modelsummary(models, coef_omit = "^(?!.*ei|.*pt)", omit = ".*")
+#' modelsummary(models, coef_omit = "^(?!.*ei|.*pt)", gof_omit = ".*")
 #' 
 #' # shape: multinomial model
 #' library(nnet)