+++ NEWS.md
@@ -1,5 +1,6 @@
-# modelsummary 1.1.0.9000
+# modelsummary 1.1.0.9001
 
+* `coef_omit` accepts numeric indices.
 * `datasummary_skim`: Increased maximum number of variables to 250.
 * Quarto notebooks compile to Word and Markdown automatically.